pyproject.toml
[project]
name = "llm-safety-eval"
version = "0.1.0"
description = "Framework for evaluating adversarial prompting and safety boundary failures in LLMs"
authors = [{ name = "Research Team", email = "example@example.com" }]
requires-python = ">=3.11,<3.13"
readme = "README.md"
dependencies = [
  "typer[all]>=0.12.3",
  "pydantic>=2.6.0",
  "pandas>=2.2.0",
  "httpx>=0.27.0",
  "jsonschema>=4.21.0",
  "matplotlib>=3.8.0",
  "pyyaml>=6.0",
  "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.3.0",
  "black>=24.1.1",
  "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
extend-select = ["B", "C4", "I", "N", "T20"]
target-version = "py311"
ignore = ["D203", "D212"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
